Remove background service
